rename sitemap-item
