+++ lib/sitemap-stream.ts
@@ -7,7 +7,7 @@ import {
   Writable,
 } from 'stream';
 import { validateSMIOptions, normalizeURL } from './utils';
-import { SitemapItemStream } from './sitemap-item';
+import { SitemapItemStream } from './sitemap-item-stream';
 export const preamble =
   '<?xml version="1.0" encoding="UTF-8"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">';
 export const closetag = '</urlset>';
